refactored dataloader
